fix depot save when using inmemory db

+++ src/NeoServer.Data/Repositories/PlayerDepotItemRepositoryNeo.cs
@@ -26,10 +26,19 @@ namespace NeoServer.Data.Repositories
 
         public async Task DeleteAll(uint playerId)
         {
+            if (!Context.Database.IsRelational())
+            {
+                var items = Context.PlayerDepotItems.Where(x => x.PlayerId == playerId);
+                foreach (var item in items)
+                {
+                    await Delete(item);
+                }
+                return;
+            }
             await Context.Database.ExecuteSqlRawAsync($"delete from player_depot_items where player_id = {playerId}");
         }
 
-    
+
         #endregion
     }
 }
